+++ internal/domain/usecase/get_experiment_channels_use_case.go
@@ -1,0 +1,11 @@
+package usecase
+
+import (
+	"context"
+
+	"github.com/kshmirko/lidar-platform-go/internal/domain/entity"
+)
+
+type GetExperimentChannelsUseCase interface {
+	Execute(ctx context.Context, id uint) ([]entity.ExperimentChannel, error)
+}